refactor(bridge): move event handlers to listening section
